pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "cryptography (>=45.0.4,<46.0.0)",
    "pycountry (>=24.6.1,<25.0.0)",
    "pycountry-convert (>=0.7.2,<0.8.0)",
    "xlwings (>=0.33.15,<0.34.0)",
    "logtail-python (>=0.3.3,<0.4.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
